compaction_manager: switch to table_state for mapping of compaction_state
manager stores a state for each table. As we're transitioning towards
table_state, the mapping of a table to compaction state will now use
table_state ptr as key. table_state ptr is stable and its lifetime
is the same as table.

we're temporarily adding a ptr to compaction_state, as there's lots
of dependency on replica::table, but we'll get rid of it once
we complete the transition.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>