service/raft: raft_group0: make some calls abortable
There are some calls to `modify_config` which should react to aborts
(e.g. when we shutdown Scylla).

There are also calls to `send_group0_modify_config` which should
probably also react to aborts, but the functions don't take
an abort_source parameter. This is fixable but I left TODOs for now.