gossiper: apply updates in batches
Currently the gossiper applies each message as it lands. The apply
process splits the message into per-endpoint states and applies these
states individually. This has a limited concurrency of 100, further
updates are queued on the semaphore. In practice this concurrency can be
reduced further as each update locks the endpoint it is updating, so if
there are multiple concurrent updates for the same endpoint, only one is
processed, the rest waits on the lock.
Updates are replicated to other shards as they are applied. This has
negligible impact usually. But as the number of nodes (and therefore the
number of messages to process) and number of shards scales up, the
chance of some shard lagging behind with applying the updates increases.
Applying the updates on remote shards is very little amount of work, but
the entire process is very sensitive to latency: the latency of applying
the message is that of the slowest shard replicating the update. It is
particularly bad if one of the shards is loaded with a process that is
violating its task-quota for an extended period of time. This means that
even though the gossip scheduling group still gets it runtime, it will
get it as little as just a few times per second, meaning that only a few
endpoint updates can be replicated per second. In a large cluster where
a single message can contain dozens of endpoint state updates, this can
lead to unprocessed messages queueing up unbounded, leading to OOM.

This patch attempts to alleviate this problem by squashing all
unprocessed updates into a single staging state. Applying the state is
moved into a separate loop, which applies the entire state in
iterations. While the loop is in an iteration, any incoming updates are
applied to the staging state, after which the loop is notified that
there is new state to process.

This patch is a very rough sketch of the idea. Many things are not
addressed:
* Intricacies of applying states;
* Error handling;
* Error propagation;
* Waiting for individual messages;

The goal of this patch is to start a discussion on this proposal based
on real code, without expending large amount of effort on correctness
and polishing on an idea that is potentially fundamentally flawed.