utils/histogram.hh: should_sample should use a bitmask
This patch fixes a bug in should_sample that uses its bitmask
incorrectly.

basic_ihistogram has a feature that allows it to sample values instead
of taking a timer each time.

To decide if it should sample or not, it uses a bitmask. The bitmask
is of the form 2^n-1, which means 1 out of 2^n will be sampled.

For example, if the mask is 0x1 (2^2-1) 1 out of 2 will be sampled.
If the mask is 0x7 (2^3-1) 1 out of 8 will be sampled.

There was a bug in the should_sampled() method.
The correct form is (value&mask) == mask

Ref #2747
It does not solve all of #2747, just the bug part of it.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>