time_window_compaction_strategy: get_sstables_for_compaction: improve debug messages
Print the compaction_strategy `this` pointer
so we can distinguish between different instance of the
compaction_strategy object (some code paths copy it and
some may instantiate a brand new compaction_strategy object).

The motivation is detecting when the side effects of this function are
applied on the "master" instance, stored in the table shard.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>