cql3: Don't create single element conjunction
In case the expression is empty and we want to merge it
with a new restriction we can just set the expression
to the new restriction.

Later this will make it easier to distinguish which case
of multi column restrictions are we dealing with.

IN and EQ can only have a single binary operator,
but slice might have two.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>