service: storage_proxy: introduce `filter_for_query` helpers
The helpers pass a pointer to `_gossiper` to `db::filter_for_query`
and forward all other arguments. This is done to reduce the number of
different functions that directly access `_gossiper` in `storage_proxy`.

Remove one overload of `db::filter_for_query`, it is replaced with
an overload of `storage_proxy::filter_for_query`.