doc, alternator: document the possibility of write reordering
In issue #10966, a user noticed that Alternator writes may be reordered
(a later write to an item is ignored with the earlier write to the same
item "winning") if Scylla nodes do not have synchronized time and if
always_use_lwt write isolation mode is not used.

In this patch I add to docs/alternator/compatibility.md a section about
this issue, what causes it, and how to solve or at least mitigate it.

Fixes #10966

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

Closes #11094