hintedhandoff_additional_test, make it work with raft
Schema changes require a quorum of nodes, which means that a cluster with two nodes and one down node cannot apply the schema change.

The general approach to fixing dtests for raft
is to go over all of them one by one and add --experimental_features raft.
Then, when all of them are fixed, we'll change the default and
make raft non-experimental.

Fix: scylladb/scylla-dtest#2753