locator: snitch_base: remove `reload_gossiper_state` from the interface
The function was used only by a single snitch implementation:
`gossiping_property_file_snitch`. Make it a non-virtual private member
of the latter.

Introduce `i_endpoint_snitch& get()` member function in `snitch_ptr`,
use it to down-cast the interface to the concrete class in order to
access the private member when performing cross-shard ops.